Move exceptions from exceptions.py to errors.py
