--- conda_verify/exceptions.py
@@ -1,6 +1,0 @@
-class RecipeError(Exception):
-    pass
-
-
-class PackageError(Exception):
-    pass
